requirements.txt
aerich==0.7.2
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.2
certifi==2024.2.2
click==8.1.7
cloudinary==1.38.0
dictdiffer==0.9.0
dnspython==2.5.0
email-validator==2.1.0.post1
Faker==23.2.1
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
iso8601==1.1.0
Jinja2==3.1.3
litestar==2.6.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mirakuru==2.5.2
msgspec==0.18.6
multidict==6.0.5
packaging==23.2
passlib==1.7.4
pluggy==1.4.0
polyfactory==2.14.1
port-for==0.7.2
psutil==5.9.8
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg-pool==3.2.1
pydantic==2.6.1
pydantic-openapi-schema==1.5.1
pydantic-settings==2.2.0
pydantic_core==2.16.2
Pygments==2.17.2
PyJWT==2.8.0
pypika-tortoise==0.1.6
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-postgresql==5.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
rich==13.7.0
rich-click==1.7.3
six==1.16.0
sniffio==1.3.0
text-unidecode==1.3
tomlkit==0.12.3
tortoise-orm==0.20.0
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.1
websockets==12.0
